Filter moved to web[dot]xml... for filter order preservation
