+++ src/main/webapp/WEB-INF/web.xml
@@ -21,5 +21,32 @@
         </user-data-constraint>
     </security-constraint>
 
+    <filter>
+        <filter-name>AuthorizationFilter</filter-name>
+        <filter-class>in.project.tasktracker.Filters.AuthorizationFilter</filter-class>
+    </filter>
+    <filter-mapping>
+        <filter-name>AuthorizationFilter</filter-name>
+        <url-pattern>/</url-pattern>
+        <url-pattern>/newTask</url-pattern>
+        <url-pattern>/updateTaskStatus</url-pattern>
+        <url-pattern>/deleteTask</url-pattern>
+        <url-pattern>/profile</url-pattern>
+        <url-pattern>/logout</url-pattern>
+        <url-pattern>/upload-profile-pic</url-pattern>
+        <url-pattern>/delete-profile-pic</url-pattern>
+        <url-pattern>/updateProfile</url-pattern>
+    </filter-mapping>
+
+    <filter>
+        <filter-name>PasswordFilter</filter-name>
+        <filter-class>in.project.tasktracker.Filters.PasswordFilter</filter-class>
+    </filter>
+    <filter-mapping>
+        <filter-name>PasswordFilter</filter-name>
+        <url-pattern>/signup</url-pattern>
+        <url-pattern>/updateProfile</url-pattern>
+    </filter-mapping>
+
 
 </web-app>
